requirements.txt
safetensors==0.5.3
requests==2.32.3
PyYAML==6.0.2
scipy>=1.11
datasets==3.1.0
numpy>=1.26
accelerate==1.0.1
biopython==1.83

# REMOVE deepspeed — unsupported on macOS
# deepspeed==0.15.4

torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

tqdm==4.67.1
transformers==4.46.3

# macOS-compatible FAISS
faiss-cpu==1.13.0

scikit-learn==1.3.2
openai==1.68.2

# REMOVE Triton (never works on macOS)
# triton==3.0.0

rouge==1.0.1
rouge_score==0.1.2
six==1.17.0
tokenizers==0.20.3

torchmetrics
seaborn==0.13.2
sentencepiece==0.2.0
huggingface-hub==0.29.3
peft==0.13.2
pyarrow==17.0.0
nltk==3.9.1
pandas==2.0.3
lmdb==1.6.2
pillow==10.4.0
einops==0.8.1
